change : Tests 코드 분할
